Bump pinned numpy to 1.10.4
We need to do this to work around an issue with mkl packages shipped with
older version of numpy. See:

  https://github.com/ContinuumIO/anaconda-issues/issues/698#issuecomment-204806498

for more. Using numpy 1.9.x, the code would build with mkl but ip-diffim
would fail tests because mkl couldn't find all of its libraries and
therefore symbols (would terminate with `python: symbol lookup error`).  See
the above issue for more.